fix(build): reorganize server modules to correctly split client/ server
