+++ app/routes/explore.forgot.tsx
@@ -1,6 +1,17 @@
 import * as React from "react";
 import { useTranslation } from "react-i18next";
-import  { type ActionFunctionArgs, type LoaderFunctionArgs, type MetaFunction, data, redirect, Form, Link, useActionData, useNavigation, useSearchParams  } from "react-router";
+import {
+  type ActionFunctionArgs,
+  type LoaderFunctionArgs,
+  type MetaFunction,
+  data,
+  redirect,
+  Form,
+  Link,
+  useActionData,
+  useNavigation,
+  useSearchParams,
+} from "react-router";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import ErrorMessage from "~/components/error-message";
@@ -14,7 +25,7 @@ import {
   CardHeader,
   CardTitle,
 } from "~/components/ui/card";
-import { validateEmail } from "~/utils.server";
+import { validateEmail } from "~/utils";
 import { getUserId } from "~/utils/session.server";
 
 export async function loader({ request }: LoaderFunctionArgs) {
@@ -69,7 +80,7 @@ export async function action({ request }: ActionFunctionArgs) {
       {
         errors: { email: "An error occurred. Please try again later." },
         success: false,
-        error: error
+        error: error,
       },
       { status: 500 },
     );
