+++ app/routes/explore.login.tsx
@@ -1,6 +1,17 @@
 import * as React from "react";
 import { useTranslation } from "react-i18next";
-import  { type ActionFunctionArgs, type LoaderFunctionArgs, type MetaFunction, data, redirect, Form, Link, useActionData, useNavigation, useSearchParams  } from "react-router";
+import {
+  type ActionFunctionArgs,
+  type LoaderFunctionArgs,
+  type MetaFunction,
+  data,
+  redirect,
+  Form,
+  Link,
+  useActionData,
+  useNavigation,
+  useSearchParams,
+} from "react-router";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import ErrorMessage from "~/components/error-message";
@@ -16,7 +27,7 @@ import {
 } from "~/components/ui/card";
 import { Checkbox } from "~/components/ui/checkbox";
 import { verifyLogin } from "~/models/user.server";
-import { safeRedirect, validateEmail } from "~/utils.server";
+import { safeRedirect, validateEmail } from "~/utils";
 import { createUserSession, getUserId } from "~/utils/session.server";
 
 export async function loader({ request }: LoaderFunctionArgs) {
