+++ app/routes/join.tsx
@@ -1,13 +1,22 @@
 import i18next from "app/i18next.server";
 import * as React from "react";
-import  { type ActionFunctionArgs, type LoaderFunctionArgs, type MetaFunction, data, redirect , Form, Link, useActionData, useSearchParams  } from "react-router";
+import {
+  type ActionFunctionArgs,
+  type LoaderFunctionArgs,
+  type MetaFunction,
+  data,
+  redirect,
+  Form,
+  Link,
+  useActionData,
+  useSearchParams,
+} from "react-router";
 
 import ErrorMessage from "~/components/error-message";
 import { getProfileByUsername } from "~/models/profile.server";
 import { createUser, getUserByEmail } from "~/models/user.server";
-import { safeRedirect, validateEmail, validateName } from "~/utils.server";
+import { safeRedirect, validateEmail, validateName } from "~/utils";
 import { createUserSession, getUserId } from "~/utils/session.server";
-
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const userId = await getUserId(request);
