+++ app/routes/settings.password.tsx
@@ -1,5 +1,12 @@
 import { useEffect, useRef } from "react";
-import  { type ActionFunctionArgs, type LoaderFunctionArgs, data, redirect , Form, useActionData  } from "react-router";
+import {
+  type ActionFunctionArgs,
+  type LoaderFunctionArgs,
+  data,
+  redirect,
+  Form,
+  useActionData,
+} from "react-router";
 import invariant from "tiny-invariant";
 import { useToast } from "@/components/ui/use-toast";
 import ErrorMessage from "~/components/error-message";
@@ -15,7 +22,7 @@ import {
 import { Input } from "~/components/ui/input";
 import { Label } from "~/components/ui/label";
 import { updateUserPassword, verifyLogin } from "~/models/user.server";
-import { validatePassLength, validatePassType } from "~/utils.server";
+import { validatePassLength, validatePassType } from "~/utils";
 import { getUserEmail, getUserId } from "~/utils/session.server";
 
 //*****************************************************
