+++ tests/utils.spec.ts
@@ -1,4 +1,4 @@
-import { validateEmail } from "~/utils.server"
+import { validateEmail } from "~/utils";
 
 test("validateEmail returns false for non-emails", () => {
   expect(validateEmail(undefined)).toBe(false);
